# Swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "College Management System",
    "contact": {}
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/admin/activateAccount": {
      "post": {
        "summary": "Activate Account",
        "tags": [
          "Admins"
        ],
        "operationId": "activateAccount",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/activateAccountRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/admin/forgetPassword": {
      "post": {
        "summary": "Forget Password",
        "tags": [
          "Admins"
        ],
        "operationId": "forgetPassword",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/forgetPasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/verifyOtpCode": {
      "post": {
        "summary": "Verify Otp Code",
        "tags": [
          "Admins"
        ],
        "operationId": "verifyOtpCode",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/verifyOtpCodeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admin/changePassword": {
      "post": {
        "summary": "Change Password",
        "tags": [
          "Admins"
        ],
        "operationId": "changePassword",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/changePasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/student/activateAccount": {
      "post": {
        "summary": "Activate Account",
        "tags": [
          "Students"
        ],
        "operationId": "activateAccount",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/activateAccountRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/student/forgetPassword": {
      "post": {
        "summary": "Forget Password",
        "tags": [
          "Students"
        ],
        "operationId": "forgetPassword",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/forgetPasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/student/verifyOtpCode": {
      "post": {
        "summary": "Verify Otp Code",
        "tags": [
          "Students"
        ],
        "operationId": "verifyOtpCode",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/verifyOtpCodeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/student/changePassword": {
      "post": {
        "summary": "Change Password",
        "tags": [
          "Students"
        ],
        "operationId": "changePassword",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/changePasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/teacher/activateAccount": {
      "post": {
        "summary": "Activate Account",
        "tags": [
          "Teachers"
        ],
        "operationId": "activateAccount",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/activateAccountRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/teacher/forgetPassword": {
      "post": {
        "summary": "Forget Password",
        "tags": [
          "Teachers"
        ],
        "operationId": "forgetPassword",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/forgetPasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/teacher/verifyOtpCode": {
      "post": {
        "summary": "Verify Otp Code",
        "tags": [
          "Teachers"
        ],
        "operationId": "verifyOtpCode",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/verifyOtpCodeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/teacher/changePassword": {
      "post": {
        "summary": "Change Password",
        "tags": [
          "Teachers"
        ],
        "operationId": "changePassword",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/changePasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "activateAccountRequest": {
      "title": "Activate Account Request",
      "example": {
        "email": "email",
        "accountActivationCode": "123456"
      },
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "accountActivationCode": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "accountActivationCode"
      ]
    },
...
    "forgetPasswordRequest": {
      "title": "Forget Password Request",
      "example": {
        "email": "email"
      },
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        }
      },
      "required": [
        "email"
      ]
    },
    "verifyOtpCodeRequest": {
      "title": "Verify Otp Code Request",
      "example": {
        "email": "email",
        "otpCode": "123456",
        "newPassword": "newpassword",
        "confirmPassword": "confirmpassword"
      },
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "otpCode": {
          "type": "string"
        },
        "newPassword": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "otpCode",
        "newPassword",
        "confirmPassword"
      ]
    },
    "changePasswordRequest": {
      "title": "Change Password Request",
      "example": {
        "email": "email",
        "oldPassword": "oldpassword",
        "newPassword": "newpassword",
        "confirmPassword": "confirmpassword"
      },
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "oldPassword": {
          "type": "string"
        },
        "newPassword": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "oldPassword",
        "newPassword",
        "confirmPassword"
      ]
    },
...
  },
...
}